delete assets func
