+++ src/init.js
@@ -2,9 +2,9 @@ import { gsap } from 'gsap';
 import { PixiPlugin } from 'gsap/PixiPlugin';
 import { Viewport } from 'pixi-viewport';
 import * as PIXI from 'pixi.js';
-import { transformManifest } from './assets/asset';
 import { firaCode } from './assets/fonts';
 import { icons } from './assets/icons';
+import { transformManifest } from './assets/utils';
 import * as viewportEvents from './events/viewport';
 import { deepMerge } from './utils/deepmerge/deepmerge';
 gsap.registerPlugin(PixiPlugin);
